This code illustrates the concept of polymorphism, where objects of different derived classes can be treated as objects of the base class but behave differently based on their specific implementations of overridden functions.
We have a base class named "Shape" with a virtual function "draw." The virtual keyword allows derived classes to override this function.

The "draw" function in the "Shape" class provides a generic message, indicating that it's drawing a shape. This message serves as a placeholder and can be customized by derived classes.

We also define a derived class named "Circle," which inherits from the "Shape" class.

In the "Circle" class, we override the "draw" function. This means that when an object of the "Circle" class calls the "draw" function, it will execute the code defined in the "Circle" class, which outputs a specific message for drawing a circle