Create encapsulation.cpp
This code defines a class called Circle that has a private member variable called radius. The __init__ method of the class initializes the radius member variable to the value passed to it. The get_area method of the class calculates and returns the area of the circle.

The main function of the code creates an instance of the Circle class and calls the get_area method to calculate the area of the circle. The area is then printed to the console.